use seperate networks to simulate ... network
  - only for the satellite
    copy inline template config direct into the PAI path (bad hack)
  - use the sign.json for replace the icinga2-master name in the Endpoint config

+++ rootfs/init/runtime/zone_watcher.sh
@@ -64,16 +64,7 @@ inotifywait \
         cp /etc/icinga2/zones.conf ${ICINGA_LIB_DIR}/backup/zones.conf
 
         log_info "we remove also the static global-templates directory"
-        if [[ -d /etc/icinga2/zones.d/global-templates ]]
-        then
-          for template in $(ls -1 /etc/icinga2/zones.d/global-templates/*)
-          do
-            name=$(basename ${template})
-            [[ -f /var/lib/icinga2/api/zones/global-templates/_etc/${base} ]] || cp -v ${template} /var/lib/icinga2/api/zones/global-templates/_etc/
-          done
-
-          rm -rf /etc/icinga2/zones.d/global-templates
-        fi
+        [[ -d /etc/icinga2/zones.d/global-templates ]] && rm -rf /etc/icinga2/zones.d/global-templates
 
         log_info "we remove also the static director-global directory"
         [[ -d /etc/icinga2/zones.d/director-global ]] && rm -rf /etc/icinga2/zones.d/director-global
